# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Industrial Contract & Supplier Lifecycle Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, pending, terminated]
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    put:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractRequest"
      responses:
        "200":
          description: Contract updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Contract deleted
  /contracts/{id}/analyze:
    post:
      operationId: analyzeContract
      tags: [contracts]
      summary: AI-analyze a contract PDF to extract dates and clauses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractAnalysis"
  /contracts/{id}/risk:
    get:
      operationId: getContractRisk
      tags: [contracts]
      summary: Get expiry risk score for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"
  /contracts/expiry-calendar:
    get:
      operationId: getExpiryCalendar
      tags: [contracts]
      summary: Get contracts grouped by expiry month for calendar view
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supplier, client]
        supplierId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, createdAt]
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
          nullable: true
        contractNumber:
          type: string
        status:
          type: string
          enum: [active, expired, pending, terminated]
        effectiveDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        autoRenewal:
          type: boolean
        autoRenewalTerms:
          type: string
          nullable: true
        value:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        riskLevel:
          type: string
          enum: [low, medium, high]
          nullable: true
        daysUntilExpiry:
          type: integer
          nullable: true
        aiExtracted:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, supplierId, contractNumber, status, autoRenewal, aiExtracted, createdAt, updatedAt]
    CreateContractRequest:
      type: object
      properties:
        title:
          type: string
        supplierId:
          type: integer
        contractNumber:
          type: string
        status:
          type: string
          enum: [active, expired, pending, terminated]
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        autoRenewal:
          type: boolean
        autoRenewalTerms:
          type: string
        value:
          type: number
        currency:
          type: string
        notes:
          type: string
      required: [title, supplierId, contractNumber]
    UpdateContractRequest:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [active, expired, pending, terminated]
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        autoRenewal:
          type: boolean
        autoRenewalTerms:
          type: string
        value:
          type: number
        currency:
          type: string
        notes:
          type: string
    ContractAnalysis:
      type: object
      properties:
        effectiveDate:
          type: string
          nullable: true
        expiryDate:
          type: string
          nullable: true
        autoRenewal:
          type: boolean
        autoRenewalTerms:
          type: string
          nullable: true
        keyTerms:
          type: array
          items:
            type: string
        confidence:
          type: number
        rawExtraction:
          type: string
      required: [autoRenewal, keyTerms, confidence, rawExtraction]
    RiskAssessment:
      type: object
      properties:
        contractId:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high]
        daysUntilExpiry:
          type: integer
          nullable: true
        score:
          type: number
        factors:
          type: array
          items:
            type: string
        recommendation:
          type: string
      required: [contractId, riskLevel, score, factors, recommendation]
    CalendarEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        contracts:
          type: array
          items:
            $ref: "#/components/schemas/ContractSummary"
      required: [date, contracts]
    ContractSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        supplierName:
          type: string
          nullable: true
        riskLevel:
          type: string
          nullable: true
        status:
          type: string
      required: [id, title, status]
...