# lib/api-zod/src/generated/types/calendarEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industrial Contract & Supplier Lifecycle Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ContractSummary } from "./contractSummary";

export interface CalendarEntry {
  date: Date;
  contracts: ContractSummary[];
}