# lib/api-zod/src/generated/types/contractSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industrial Contract & Supplier Lifecycle Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ContractSummary {
  id: number;
  title: string;
  supplierName?: string | null;
  riskLevel?: string | null;
  status: string;
}