# lib/api-zod/src/generated/types/createContractRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industrial Contract & Supplier Lifecycle Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContractRequestStatus } from "./createContractRequestStatus";

export interface CreateContractRequest {
  title: string;
  supplierId: number;
  contractNumber: string;
  status?: CreateContractRequestStatus;
  effectiveDate?: Date;
  expiryDate?: Date;
  autoRenewal?: boolean;
  autoRenewalTerms?: string;
  value?: number;
  currency?: string;
  notes?: string;
}